social_media_api/requirements.txt
altgraph @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
asgiref==3.9.1
dj-database-url==3.0.1
Django==4.2.23
django-filter==25.1
django-taggit==6.1.0
djangorestframework==3.16.1
future @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
gunicorn==23.0.0
macholib @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
mysql-connector-python==9.4.0
packaging==25.0
pillow==11.3.0
python-decouple==3.8
python-dotenv==1.1.1
six @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
sqlparse==0.5.3
typing_extensions==4.15.0
whitenoise==6.11.0
